requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
cachetools==5.5.2
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
coverage==7.13.5
cryptography==46.0.5
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
dpath==2.2.0
dulwich==1.1.0
dvc==3.50.1
dvc-data==3.15.2
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
email-validator==2.3.0
entrypoints==0.4
fastapi==0.111.0
fastapi-cli==0.0.24
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gto==1.9.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
huey==2.6.0
huggingface_hub==1.7.1
hydra-core==1.3.2
identify==2.6.18
idna==3.11
importlib_metadata==7.2.1
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaggle==2.0.0
kiwisolver==1.5.0
kombu==5.6.2
lightgbm==4.3.0
loguru==0.7.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlflow==2.13.0
multidict==6.7.1
mypy==1.10.0
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==1.26.4
omegaconf==2.3.0
orjson==3.11.7
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
patsy==1.0.2
pillow==12.1.1
pluggy==1.6.0
pre-commit==3.7.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pyarrow==15.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pydot==4.0.1
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.2
pytest==8.2.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.3
querystring-parser==1.2.4
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
ruamel.yaml==0.19.1
ruff==0.4.4
scikit-learn==1.4.2
scipy==1.17.1
scmrepo==3.6.1
semver==3.0.4
setuptools==82.0.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.5
sqltrie==0.11.2
starlette==0.37.2
statsmodels==0.14.2
tabulate==0.10.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.39.1
voluptuous==0.16.0
waitress==3.0.2
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.6
yarl==1.23.0
zc.lockfile==4.0
zipp==3.23.0
